requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
pypdf==3.17.0
pytesseract==0.3.10
pillow==10.1.0
markdown==3.5
beautifulsoup4==4.12.2
pygments==2.16.1
pymupdf==1.21.1
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.335
langchain-openai
openai
opencv-python==4.8.1.78
transformers==4.35.2
numpy==1.26.1
tqdm==4.66.1
pytest==7.4.3
httpx==0.25.1